Refactor S3 upload functionality: move uploadBlob to utils, simplify upload process in uploadToS3
